requirements.txt
robin_stocks
python-dotenv
discord.py
black
money
babel